bins: intial CPU binding support
add bindings for `processor_bind()`, the mechanism to make this happen.
then, in propolis-standalone, allow a choice of how binding should be
done: allow vCPUs to roam with `cpu_binding = "any"` or pin vCPUs to
CPUs 1:1 from the last processor downwards with
`cpu_binding = "from-last"`.

in `propolis-server`, pick between these two strategies automatically
based on how the vCPU count compares to the number of available
processors. a VM that would use more than half of available
processors is bound in the same way as `propolis-standalone` with
`cpu_binding = "from-last"`, but otherwise continues to have unbound
vCPUs.

in the future, Nexus will dictate `propolis-server`'s vCPU bindings,
but exactly how is a bit of an open question.